+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.2.0 (2016-08-2)
+------------------
+
+* Change the way how to distinguish permission-related method from others.
+* Now you should use ``object_permission`` decorator instead of ``can_`` prefix to define permission methods.
+
+
 0.1.0 (2016-07-19)
 ------------------